LPS-111561 Isolate List Item Style selector
